Creating MainReportsScreen style file
